+++ src/compas/datastructures/assembly/assembly.py
@@ -116,6 +116,7 @@ class Assembly(Datastructure):
         part.key = key
         self._parts[part.guid] = part.key
         return key
+    
 
     def add_connection(self, a, b, **kwargs):
         """Add a connection between two parts.
@@ -146,6 +147,40 @@ class Assembly(Datastructure):
         if not self.graph.has_node(a.key) or not self.graph.has_node(b.key):
             raise AssemblyError(error_msg)
         return self.graph.add_edge(a.key, b.key, **kwargs)
+    
+
+    def delete_part(self, part):
+        """Remove a part  from the assembly.
+
+        Parameters
+        ----------
+        part : :class:`compas.datastructures.Part`
+            The part to add.
+
+        Returns
+        -------
+        None
+
+        """
+        del self._parts[part.guid]
+        self.graph.delete_node(key=part.key)
+
+    
+    def delete_connection(self, edge):
+        """Delete a connection between two parts.
+
+        Parameters
+        ----------
+        edge : :class:`compas.datastructures.Part`
+            The part to add.
+
+        Returns
+        -------
+        None
+
+        """
+        self.graph.delete_edge(edge=edge)
+   
 
     def parts(self):
         """The parts of the assembly.